# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tracklet API Documentation",
        "description": "Multi-Organization Management System API - Complete API documentation for mobile and web developers",
        "contact": {
            "email": "support@tracklet.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "API Server (uses current host)"
        }
    ],
    "paths": {
        "/api/assets": {
            "get": {
                "tags": [
                    "Assets"
                ],
                "summary": "List assets",
                "operationId": "7ec31142a2a1e2a05ac095eec4473640",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "active",
                                "in_repair",
                                "retired"
                            ]
                        }
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "assigned_to_user_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of assets"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Assets"
                ],
                "summary": "Create asset",
                "operationId": "009a09a3ccd2d061f0a7f2e7b40696a9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "category",
                                    "purchase_date",
                                    "purchase_price"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Dell Laptop"
                                    },
                                    "category": {
                                        "type": "string",
                                        "example": "Electronics"
                                    },
                                    "purchase_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-01-15"
                                    },
                                    "purchase_price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "1200.00"
                                    },
                                    "vendor": {
                                        "type": "string",
                                        "example": "Dell Inc"
                                    },
                                    "warranty_expiry": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "assigned_to_user_id": {
                                        "type": "integer"
                                    },
                                    "assigned_to_location": {
                                        "type": "string",
                                        "example": "Room 101"
                                    },
                                    "serial_number": {
                                        "type": "string"
                                    },
                                    "model_number": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Asset created successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/assets/{id}": {
            "get": {
                "tags": [
                    "Assets"
                ],
                "summary": "Get asset",
                "operationId": "c350c4b3ce25b74af75712ce0e2675ba",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Asset details"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Assets"
                ],
                "summary": "Update asset",
                "operationId": "b5e1e1d7b4b147885862b3597820d9bf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "category",
                                    "purchase_date",
                                    "purchase_price",
                                    "status"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "category": {
                                        "type": "string"
                                    },
                                    "purchase_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "purchase_price": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "active",
                                            "in_repair",
                                            "retired"
                                        ]
                                    },
                                    "status_change_reason": {
                                        "type": "string"
                                    },
                                    "assigned_to_user_id": {
                                        "type": "integer"
                                    },
                                    "assigned_to_location": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Asset updated successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Assets"
                ],
                "summary": "Delete asset",
                "operationId": "e3fab23c9ce87df5bff3d910636bcafe",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Asset deleted successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/assets/{id}/movement": {
            "post": {
                "tags": [
                    "Assets"
                ],
                "summary": "Log asset movement",
                "operationId": "8af2d252cf20020c80b34d1cf82fa539",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "movement_date",
                                    "movement_type"
                                ],
                                "properties": {
                                    "movement_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-11-28"
                                    },
                                    "movement_type": {
                                        "type": "string",
                                        "enum": [
                                            "assignment",
                                            "location_change",
                                            "return",
                                            "other"
                                        ],
                                        "example": "assignment"
                                    },
                                    "from_user_id": {
                                        "type": "integer"
                                    },
                                    "from_location": {
                                        "type": "string"
                                    },
                                    "to_user_id": {
                                        "type": "integer"
                                    },
                                    "to_location": {
                                        "type": "string",
                                        "example": "Room 102"
                                    },
                                    "reason": {
                                        "type": "string",
                                        "example": "Department transfer"
                                    },
                                    "notes": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Movement logged successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}